Make the ring lifecycler use the same code to pick the KSStore implementation as the ring does.
Also, remove the two instances of KVClient mock hooks.
